psc_demo/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
pg8000
cloud-sql-python-connector
